work on typechecking a bit
* fix subtyping error
* inference for application

Unfortunately there's no way to test inference for application because
we can't infer the type of lambdas and there's no way to do type
ascription.  We should probably add a syntactic form for type
ascription, or perhaps add a syntactic form for lambdas whose binders
are annotated with a type.