actually, empty case should not be allowed
An empty case is *not* the right thing to eliminate Void (the comment
deleted by this commit was wrong).  In Haskell, say, you have to do a
case *on* some expression.  So if v :: Void, then

  case v of { }

is how you eliminate v.  But in our context an "empty case" is
really *literally* empty, with nothing even to match on, so it makes
little sense.