bug fix: closures have to have variables of the right type!
Otherwise the variables don't match and don't bind the right things!
Also it was silly to unbind and rebind a term just to stick it in a
closure.

This was causing all sorts of trouble, e.g. recursive functions were
just looping forever since the variable in the closure never matched any
of the variables in the body, so the variables in the body always just
had the same value as the very first time the function was called.