More efficient list parsing
Previously, deeply nested lists caused very bad performance: to parse
[[[[[3]]]]], it would first try parsing "[[[[3]]]]" as a list
comprehension, which it did by first trying to parse a term followed
by |: to parse [[[[3]]]] as a term, it would try parsing [[[3]]] as a
list comprehension ... and so on.  But after this failed and it
decided to parse the outer list as a literal list, it would re-parse
the first term in the list, and the whole process would be repeated,
resulting in n + (n-1) + (n-2) + (n-3) ... = O(n^2) recursive calls to
parseList for nested list of depth n.  After this change, parsing
lists should be O(n).