subtyping study: another bug fix in simplify step
'lfresh' does not prevent the use of the same name again later; for
that one needs to explicitly use 'avoid' with a continuation.  The
only reason I was using LFresh instead of Fresh was to be able to
start out the computation with an avoid set consisting of any free
type variables in the constraints.  Instead, I now find the largest
integer suffix used on any free type variable, and use contFreshMT to
start the counter at the number one greater.