good progress on formatting messages etc.
Tests don't pass since reporting test runs comes out in the wrong
order: the test output still uses iputStrLn and friends, whereas the
"Loading..." messages and so on use the message logging framework, so
they don't get printed immediately.

Should probably add a flag so we can set the message system in
immediate or buffered mode: in immediate mode, any emitted message
will be rendered and output immediately, then discarded.  In buffered
mode, it will be saved for outputting later.  When running tests in
particular, we will want to be in immediate mode, so we can see the
progress while the tests are running (some of which may take a while
to run).