bug fix: use FreshM instead of LFreshM in TCM
LFreshM was really the wrong thing to do.  With LFreshM I forgot that
you have to explicitly manage which names are "in scope" and hence to
be avoided via the 'avoid' function.  We were not doing that at all,
and in fact I think it would be very difficult to do that correctly
given the way the type checker is structured.  The result was that the
same type variable was being chosen repeatedly, leading to unsolvable
constraints (because all the constraints applied to a single type
variable 'a' instead of having multiple different variables with
different constraints).  The solution is to simply switch to FreshM,
lfresh -> fresh, lunbind -> unbind, etc.  Now all fresh type variables
are globally unique.

As an example,
```
let n = 3 in {? true if 2 divides n, false otherwise ?} :: false :: []
```
was not working before: it used the same type variable 'a' as both the
type of the list elements as well as the type of the arguments to
'divides'.  So 'a' had to be a supertype of B as well as N; the
constraint solver (correctly) complained that there is no such type.
The above example works correctly now.