use Embed Type instead of Type in patterns
This was causing the weird bug: any type variables which ended up in a
type annotation in a pattern were being treated as binders.  The
solution is to wrap all the type annotations in patterns with Embed.
Thankfully, due to the magic of bidirectional pattern synonyms, none
of the other code has to change!