bug fix: correctly infer type of multinomial coefficient
The second argument of an application of the `choose` operator can be
either a natural number (a binomial coefficient) or a list of natural
numbers (a multinomial coefficient).  We were still using

`check t2 TyN <|> check t2 (TyList TyN)`

to check the type of the second argument.  The problem is that if the
type of `t2` is a type variable, then `check t2 TyN` may "succeed" and
generate a constraint that its type is `TyN`, even if that constraint
can't be satisfied.  In that case we never try `check t2 (TyList TyN)`
and just fail later when the constraint solver is called.

For example, `let x = [1,2,2] in 5 choose x` should typecheck, but it
was failing with a solver error.  The problem is that the type of `x`
is not statically known at the point where `choose` is being checked.

The solution is to move the 'or' into the constraint itself,
generating a `COr` constraint so the solver can try both
possibilities.