parse patterns via terms
Get rid of dedicated pattern parsers: instead, parse a term and
then attempt to convert it to a pattern.

This is in preparation for (1) arithmetic patterns, which muddy the
waters by allowing terms and term-like things in patterns; parsing
patterns this way means we can cleanly allow this without difficulties
with ambiguous parses; (2) allowing pattern binders in comprehensions:
this way we can parse a term, check whether it is followed by <-/in,
and try to convert to a pattern if so, without needing to backtrack.