make new functions lub, cNeg and use them
Increases the types we can directly compute in inference mode, instead
of always generating a fresh type variable with constraints.

This mitigates the initial example in #113, though the second example
is still slow; to mitigate that we need a special case for cExp ty1
TyZ (see the comment in the code).