parse prims as identifiers, and turn them into prims later
Now, during typechecking, if we encounter an unbound variable we check
whether it is a recognized name of a prim, and if so, we turn it into
a prim and continue type checking it.

This means that users are free to use primitive names as local
variables; they will just shadow the prims.  This seems much better
than having a ton of reserved words.

There were also some problems with parsing prims that were prefixes of
other prims; this definitely solves that issue as well.