pretty-printing for standalone operators
For now, assumes that all standalone operators should be printed as
exactly that.  Eventually, all operators will be represented as the
application of a "standalone" operator to some arguments; then the
pretty-printer will have to do a bit more matching to see whether it
has an application that should be handled specially.