fix (modules): load stdlib modules before local
Ultimately we should have some sort of warning when there are multiple files which
could potentially be the target of an 'import' statement.  But for now let's order
the resolution so stdlib modules will be picked up before local ones.  e.g.
if you write 'import list' you will definitely get list.disco from the stdlib,
regardless of whether there is a file named list.disco in the same directory as you.