fix loading vs importing
It used to get confused when using :load on a file which happened to
have the same name as a module in the standard library.  Now :load
will only explicitly load the file in the location you name, whereas
'import' will look in the current directory as well as the standard
library.