fix loading vs importing, for real this time
Oops, the previous commit was not correct.  Having a simple `Maybe
FilePath` allowing a choice between (1) loading from a specific
directory or (2) loading from the CWD or stdlib, was not expressive
enough.  The problem is that sometimes we may want to also (3) load
from a specific directory or the stdlib; in particular this happens if
we `:load` a module from a specific directory and then need to
recursively process its `import`s, which should be loaded either from
the same directory as the original file (*not* the CWD!), *or* from
the stdlib.

This commit adds a new `Resolver` data type which specifies where we
should look for modules.