get rid of TUn and TBin AST nodes completely
Unary and binary operations were getting desugared into applications
of operator primitives, but still initially parsed as separate special
TUn and TBin nodes.  This patch gets rid of the special TUn and TBin
nodes and now parses unary and binary operators directly as a
TApp (TPrim ...) ...  TUn and TBin are now pattern synonyms, so a lot
of code that used them does not have to change.

This is a preliminary step towards #185 and #136.