Make left and right primitive functions instead of syntax (#291)
`left` and `right` are now primitive functions instead of special syntax.  This means in particular that `left` and `right` can be used independently, e.g. to ask for their type, or to pass them to a higher-order function such as `each`.

Cleaned up/clarified some related comments, and deleted a bunch of code having to do with `TInj` nodes.