Get rid of Cons in values and patterns, switch to Unit+Inj+Pair
The `Cons` constructor of various AST types, containing an Int and a
list of sub-things, was not pulling its weight at all:

  - The Int was always either 0 or 1
  - The list always had exactly 0, 1, or 2 things in it
  - It was wastefully used for both sums (in which case the list has exactly 1
    thing) and products (in which case the Int is always 0).

This commit changes things so that we now have separate Unit, Inj
Side, and Pair constructors. It makes things a tiny bit more
complicated with lists, since now e.g. a list cons (::) consists of an
Inj with a nested Pair, and we have to be careful to reduce both when
doing WHNF. But it also paves the way for making lists not quite
so magical; see https://github.com/disco-lang/disco/issues/281 .

I was kind of hoping it might speed things up, with e.g. fewer tiny
useless lists floating around, but it didn't seem to have any
appreciable affect on the runtime of the test suite.

Also add a convenience function `vint` which is like `vnum` but
operates on an `Integer` (which is very common).