assume N subtype when solving for var with no base type sub- or supertypes
Fixes #192.

Simply picking a base type that satisfies the var's sort doesn't
work. Instead, assume that any situation in which we (1) don't have
any base types as sub- or supertypes and (2) do have nontrivial sorts
must involve numeric types. Thus, treat N as a subtype of the variable
and proceed the same as the case where we do have base types as
subtypes.