WIP: process REPL input as a module
Running into a lot more issues with not being able to easily dispatch
local capabilities.  Before working on this further it's time to bite
the bullet and switch from 'capability' to something like 'polysemy'
or 'fused-effects'.  Once that is done, we can merge that refactoring
into this branch and continue.