simplify Core language: only allow single applications
There's no particular reason to have applications of something to
multiple arguments.  Ultimately I also want to remove lambdas with
multiple arguments, but that's a bit tricker due to stuff with "deep
unbinding" of forall and exists quantifiers.  I suppose we have to
compile those to a single lambda taking a product.