potentially return multiple types when doing type inference
Mostly works, but still some combinatorial explosion problems.  I
think it should be ameliorated by keeping track of variable polarities
and choosing solutions accordingly.  We only need to generate multiple
solutions for variables that occur in both positive and negative
positions.

Write a function to go through a type and return all its type
variables but with a polarity for each.  Then pass that in when
calling the solver with a bunch of constraints.