Automatically report LHS and RHS values for comparisons used in tests (#428)
Technically, this is achieved by desugaring boolean comparison operators inside a `Prop` into operators wrapped in `Should`, which triggers the testing framework to report the value on both sides of the operator.  This means (1) there is no longer any needed distinction at the surface language level between `==` and `=!=` (indeed, the latter is now a syntax error), and it also generalizes to work with any comparison operator, not just `==` and `<`.

Also generally improve test result reporting.

Closes #363 .